server-side calls: rename cmk/base/config_generation.py
